Remove `address` param from `grafana_organization`
Organization Info (ie. its address) was removed from the Grafana UI in
https://github.com/grafana/grafana/commit/21a7f572482180bd3ccf6bb47eebf8f500cb6e4c

According to
https://grafana.com/docs/grafana/v8.4/http_api/org/#update-organization
the API was not fully implemented, so I'm unconvinced this ever worked.

The property `id` is also removed as it's internal to grafana and not
something you could ever update.

Updating a `grafana_organization` never worked as `save_organization`
only implemented POSTs, (and would actually explode if you tried to
update `address`)