Fix performance of `grafana_user`
Previously, the number of API calls to look up the details of a single
user was equal to `1 + the number of existing users`.  This made
managing multiple users exponentially slow.

The way the types in this module are written, (with the Grafana API
user credentials being parameters of each resource), we can't implement
`prefetch` to fetch all users just once; but we can change the
implementation to directly lookup the user we're trying to manage.
This is _much_ faster!