Updated the link to the toml gem
The original link points to the toml project on github, this isn't a whole lot of help to someone who need to get the gem. So I've updated the link to point to https://rubygems.org/gems/toml/ which should help people working in air gapped environments download the correct gem and get the module up and working quickly.